Remove additional console logs from opt-out and CCLF imports (#976)
## 🎫 Ticket

N/A

## 🛠 Changes

- Remove extraneous console logs

## ℹ️ Context

These logs were duplicated so that they would show up in Jenkins. Now
that we are running these imports in lambdas, we can remove them.

These logs are currently showing up in the "bad logs" area in the splunk
dashboard since they don't have a "level" associated with them.

## 🧪 Validation

Unit Testing.