BCDA-8282: Update Makefile for bcda-app (#981)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8282

## 🛠 Changes

<!-- What was added, updated, or removed in this PR? -->
Added makefile commands for build-api, publish-api, build-worker, and
publish-worker

## ℹ️ Context

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->
In order to deploy BCDA on ECS, it is necessary to publish container
images to an ECR repository. This PR creates the Makefile commands to
enable that.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->
Dependent upon milestone 1 of the fargate migration completing, but the
steps work if pointed to the bcda-ssas repo.