BCDA-8302: Improve memory usage for CCLF Imports (#985)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8302

## 🛠 Changes

I apologize in advance for the size of this PR 🥲 

### Core Changes

1. **Move CCLF0-based validation logic inline to avoid opening and
parsing CCLF8 file twice**

The `validate` function has been removed, and the logic for verifying
the maxRecordLength and expected record count is now part of the actual
CCLF8 import flow.

2. **Keep zip files open and pass CCLF0 and CCLF8 file references into
`importCCLF0` and `importCCLF8` to avoid having to re-download**
 
When we call LoadCclfFiles, we are now searching for and maintaining
references to the CCLF0 and CCLF8 files for use by the import functions
later on.

3. **Pre-allocate download buffers by retrieving HEAD information before
downloading files**

This greatly reduces memory usage for S3 downloads and improves download
speed. Even with only a single file download for 500-700MB files, I was
seeing lambda memory usage up to 3GB and 5 minute downloads. Making this
change ensures we cap the memory allocation and allows the process to
take on the order of seconds instead of minutes.

### Supplementary Changes

4. **Move some archive validation logic into the initial zip file
processing.**

In particular, I am now searching for the zip file's inner CCLF0 and
CCLF8 files when loading it in `LoadCclfFiles`. This also allows us to
be more strict in the inputs for importCCLF0 and importCCLF8 and remove
null checks. This also prevents an issue we have had in the past where
we process the CCLF0 and then run into an unexpected error when calling
`importCclf8` afterwards.

1. **Organize synthetic CCLF file archives to reduce duplication and
match expected production archive formats**

Previously, the synthetic file archives split CCLF0 and CCLF8 files
across different zip files. This does not match how we receive data in
production. In addition, there were several different folders that
essentially duplicated files (e.g. `bcd`, `valid`, and `valid2` were all
essentially the same thing.)

## ℹ️ Context

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

### Behavioral Changes

This PR includes some behavioral changes due to the changes above.

1. **We no longer support CCLF files split across multiple archives.**

Previously, since we imported the entire directory at once, it was
possible to combine a CCLF0 file from one archive and a CCLF8 file from
another archive. With S3, we are triggering one lambda run for an
individual archive, so it is not possible to import across multiple
archives.

This change is now "official" in that we validate for the presence of
both a CCLF0 and CCLF8 file within a single archive before attempting to
import the archive. This was necessary so we can enforce that both the
CCLF0 and CCLF8 file references could be stored up front within
CclfZipMetadata.

2. **We fail fast if we find multiple CCLF0 or CCLF8 files within a
single archive.**

A single archive is expected to only contain one CCLF0 and one CCLF8
file. Multiple of these files was undefined behavior previously and I
believe only the latest CCLF0 + CCLF8 would be imported.

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->

Tested on `Test` environment with a ~600MB zip file. Out of memory on
the main branch after 5 minutes vs. processed with ~800MB in 23 seconds
on this branch.

![CleanShot 2024-08-10 at 15 29
55@2x](https://github.com/user-attachments/assets/2414c3c7-586f-48ac-b722-77a94a474881)