Parwinder/bcda 8414 enable linters (#1005)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8414

## 🛠 Changes

Enabled linters and fixed all open issues

## ℹ️ Context

We updated dependencies in the last sprint. During that update, we
realized our golangci file was misnamed, and we were not testing for
lint issues. The linters were disabled temporarily. This change is to
enable them again and fix outstanding issues.

## 🧪 Validation

Ran linter locally


![image](https://github.com/user-attachments/assets/7f255be2-dc9c-4853-a17e-a6f6d7930283)