Draft: CCLF lambda changes (#1008)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8405

## 🛠 Changes

<!-- What was added, updated, or removed in this PR? -->
Added the main entry to the CCLF lambda.
Added the logic flow for checking if the SQS event contains a CSV file
path, or CCLF.
Renamed CCLF lambda functions to attribution agnostic.
Added parser functions for checking CSV file.
Added testing for CSV parser functions.

## ℹ️ Context

These changes will be the entry for the lambda changes in order to make
the CCLF lambda more agonistic to an attribution lambda.

## 🧪 Validation

This PR is a part of a larger group of PR's that will come later. Some
of these entries don't function properly, such as the CSVImporter.
Therefore this PR is more-or less a check for making sure the
functionality is up to standards for the team.