update and throw OpOutcome on absence of CCLF8 (#1009)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-7514

## 🛠 Changes

Currently, we throw an error when CCLF8 files are not found using
`CCLFNotFoundError`. We also check if we are throwing this error for
`/runout` or `/all`. This has been updated to throw an
`OperationOutcome` error that is FHIR compliant. Message details
[here](https://confluence.cms.gov/display/BCDA/BCDA-7791%3A+Refine+EOY+Mode+Proposals#:~:text=Proposal%201%20%2D%20Gracefully%20handle%20ACOs%20without%20current%20attributions)

A check for `/runout` has been removed. In any scenario where we do not
have attribution, we will be throwing the same error creating
consistency between `/runout` and `/all`.

## ℹ️ Context

During E.O.Y. mode, we have a hardcoded freeze and multiple manual
activities. This is one of the initiatives to move away from E.O.Y.
mode.