BCDA-8570 Upgrade github action artifact to v4 (#1017)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8570

## 🛠 Changes

Upgrade github actions artifact-(up|down)load to v4.

## ℹ️ Context

v3 was deprecated as of Nov 30th. We were previously prevented upgrading
to v4 due to dependencies on node. PLT-338 fixed that for us and has
since been deployed.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Validation will be github actions workflows successfully passing as
normal.