Carl/bcda 8596 root cause fixes (#1023)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8596

## 🛠 Changes

Revert worker client transport options that were causing large scale
restarts. Adjust a few odds and ends with the job reworks to make things
clearer.

## ℹ️ Context

There was an incident related to jobs being stuck in an infinite loop of
sending successful requests to BFD. It seems like what was happening
under the hood was that requests were happening, then at some point
before finishing, the worker instance would restart. I am 95% sure this
is related to the transport config.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Tested and linted locally, deployed branch to test env where the same
issue was happening and is now resolved.