Epic/attribution lambda generalization (#1022)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8176

## 🛠 Changes

Convert file attribution lambda to be more generalized by allowing for
CSV files to be ingested as well.

## ℹ️ Context

This is a large epic including multiple feature branches. Overall the
goal is to convert the file attribution lambda to be more generalized by
allowing for CSV files to be ingested as well.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Each individual feature branch was PRd before merging into epic branch.
Each feature branch was also passing local and git linting and testing.

---------

Co-authored-by: austincanada <162146803+austincanada@users.noreply.github.com>
Co-authored-by: laurenkrugen-navapbc <126501259+laurenkrugen-navapbc@users.noreply.github.com>
Co-authored-by: Parwinder Bhagat <Parwinder.Bhagat@e14s.com>