Remove testing loggers (#1034)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8460

## 🛠 Changes

Removing loggers we are using to test splunk alerts.

## ℹ️ Context

Removing these loggers as we are doing a deployment. Plan is to add them
back in if need be (or more likely deploy dev from custom branch after).

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->