BCDA-8635: Migrate Platinum AMI From Jenkins to Github Workflows (#1031)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8635

## 🛠 Changes

Migration of platinum image creation from Jenkins -> Github Workflows.

## ℹ️ Context

Made as part of the larger effort to deprecate Jenkins. This requires
another PR in the ops repo that must also be merged:
https://github.com/CMSgov/bcda-ops/pull/1146

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Workflow passes and confirmed AMI is available in EC2. (latest run will
fail until ops repo is merged, but successful run can be found here:
https://github.com/CMSgov/bcda-app/actions/runs/13039689810/job/36378517357?pr=1031)

---------

Co-authored-by: Sean Fern <seanfern@navapbc.com>