BCDA-8631: Create group lambda (#1039)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8631

## 🛠 Changes

Lambda has been added so that we can call it to create groups, ad hoc.

## ℹ️ Context

Done as part of the effort to migrate away from jenkins.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Tests & workflows passing; integration test works and verified expected
database entries and lambda logs.