Fix release_type env var, cleanup (#1065)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8739

## 🛠 Changes

Fix release type env var, cleanup of descriptions

## ℹ️ Context

Slight snag on the script that sets release type, which ended up
deploying a dev build as opposed to the tag build.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Local linting and testing