BCDA-8839: Add longer sleep, add smoke-tests (#1067)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8839

## 🛠 Changes

Add smoketests to deploy process, add longer sleep, get all worker
instances and sort by launchtime for verification.

## ℹ️ Context

Smoketest work seems to be wrapping up so we might as well get that in
there now. Longer sleep (😢) as instances are taking a while to refresh.
Better getting of worker instance id to help as well.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Local linting and testing.