BCDA-8637: Input updates (#1069)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8637

## 🛠 Changes

Modified the portion that checks which ACO was selected. Evaluating
input parameters in `run` was failing, so moved it to an `if` section
under the step.

## ℹ️ Context

Could not test calling the workflow with parameters until the original
PR was merged; doing troubleshooting to make sure input parameters
working as expected.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Ran test against dev in GHA and all steps completed successfully.