BCDA-8888: Adjust deploy workflow deps, adjust slack alerts (#1072)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8888

## 🛠 Changes

Decouple smoke tests from full deploy alert. Reduce number of slack
alerts (noisy). Add build and package on push to main. Add cron
autodeploy daily.

## ℹ️ Context

Small adjustments to the deploy process we think would be useful as we
kick the tires.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Test the workflows in dev.