BCDA-8637: Only run SSAS tests in dev and test (#1073)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8637

## 🛠 Changes

Only run SSAS smoke tests in dev and test

## ℹ️ Context

These were failing in opensbx and prod. Checking back on the jenkins
scripts and we notice they are skipped there.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Local testing and linting, testing workflow in dev.