BCDA-8637: refresh attribution data GHA (#1070)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8637

## 🛠 Changes

creating github action to refresh attribution data weekly.

## ℹ️ Context

Pulled the refresh out of smoke tests to reduce number of times it runs
and so we can expand upon this for PY transitions or run it ad-hoc
without smoke tests.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Works in all envs, verified in database and multiple successful test
runs:
https://github.com/CMSgov/bcda-app/actions/runs/13996655667/job/39193385477?pr=1070