BCDA-8981: Adjust CSV file attribution (#1090)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8981

## 🛠 Changes

Add adjustments to attributing CSV metadata to account for MD TCoC and
CDAC.

## ℹ️ Context

We believe the previous process would have failed to properly find
attributions when MD TCoC and CDAC would request data.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Local linting and testing.