BCDA-8482: create config for request rate limits (#1105)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8482

## 🛠 Changes

- added a new section to the config struct -- `RateLimitContext` --
which stores the environment-specific configuration for how rate-limit
rules are applied
```
type RateLimitConfig struct {
	All  bool     `conf:"all"`  // rate-limit requests for all ACOs
	ACOs []string `conf:"acos"` // rate-limit requests for specific ACOs
}
```
- refactored the rate-limit middleware to remove environment-specific
checks, and apply rate-limit rules based on the config instead
- updated the config yml for each environment and the encrypted yml with
rate-limit rules that should reflect what is already in place for each
environment, except for sandbox, which now has a rule specific to MDTCoC

## ℹ️ Context

This ticket requires that the api block duplicate requests for a
particular ACO on a particular environment (sandbox). This means that it
needs code that executes different rules based on the environment it's
executing in. That leaves us with 3 general options:
1. **Embed this logic in the code, which will have to be
environment-aware.** I think it's generally better to leave this sort of
logic out of the code, and while we already do this very sparingly, it's
mostly in the code that already exists for the rate-limit middleware
before the changes from this PR
2. **Embed this config in the database (e.g. a table that stores ACO ids
that should be rate-limited)**
3. **Embed this config in some environment-specific config file.** This
is the approach that is taken in this PR.

## 🧪 Validation

1. Update your shared_files/decrypted/api.yml to include an aco that you
might use in testing (e.g. A9994) in the rate limit config.
2. Run `make debug-api` and attach a debugger to the running process
after it boots
3. Put a breakpoint in the `bcda_web_middleware_ratelimit.go` middleware
and ensure that the rules of the RateLimitConfig are applied as
expected.