BCDA-8962: update fhirscan to use inferno (#1108)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-8962

## 🛠 Changes

- updated FHIR Scan to run the inferno tests instead of the bdt tests. 
- removed all the temporary stuff that cloned the test kit repo locally.
- returning 0 or 1 actually means pass or fail

TODO:
 - add readme https://jira.cms.gov/browse/BCDA-9058
 - add command to the makefile https://jira.cms.gov/browse/BCDA-9059
- figure out what deployment to point the FHIRScan at
https://jira.cms.gov/browse/BCDA-9060

## ℹ️ Context

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->
part 1 created a script that had to be run manually. This ticket makes
it so you no longer have to do manual build. You just run it the same
way you used to run the BDT tests.

Right now it is pointing to the Open Sandbox, because `dev` and `test`
were having some issues (tho that was even a little flaky...). But we
can follow up to see what makes the most sense to point at

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->
Go to the actions, find fhirscan, and click 'run workflow' and select
the appropriate branch (colby/BCDA-8962)
https://github.com/CMSgov/bcda-app/actions/workflows/fhirscan.yml