BCDA-9061: update config file regex (#1114)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9061

## 🛠 Changes

Test case was added for "DA\d{4}" and config file was modified so that
"A\d{4}" is now "^A\d{4}". Latter now matches "A####" exactly rather
than a substring.

## ℹ️ Context

Regex matching for ACO names was inadvertently capturing substrings
instead of the exact match, resulting in incorrect match being returned.



<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Added test case for unit test. Configuration file needs to be updated in
all env configuration management: devs - can we verify the config file
reflects the update in lower envs?

---------

Co-authored-by: carlpartridge <carlpartridge@navapbc.com>